% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{colorize}
\alias{colorize}
\title{Colorize numerical values}
\usage{
colorize(values, resolution = 4)
}
\arguments{
\item{resolution}{Limit resolution for small values}

\item{x}{Values to normalize}
}
\value{
Colorized values
}
\description{
Colorize numerical values
}
