% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{do.community}
\alias{do.community}
\title{Do enrichment for a single community}
\usage{
do.community(
  ig,
  community,
  genesets,
  restart = 0.5,
  fdr = 1,
  pval = 1,
  plotting = FALSE
)
}
\arguments{
\item{ig}{An igraph object}

\item{community}{A community label}

\item{genesets}{A name list of genesets}

\item{restart}{Restart probability between 0-1}

\item{fdr}{Passed to hypeR}

\item{pval}{Passed to hypeR}

\item{plotting}{Passed to hypeR}
}
\value{
An igraph and hyp object
}
\description{
Do enrichment for a single community
}
