% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{do.enrichment}
\alias{do.enrichment}
\title{Do enrichment for all communities of the network}
\usage{
do.enrichment(ig, genesets, restart = 0.5, fdr = 1, pval = 1, top = 1)
}
\arguments{
\item{ig}{An igraph object}

\item{genesets}{A name list of genesets}

\item{restart}{Restart probability between 0-1}

\item{fdr}{Passed to hypeR}

\item{pval}{Passed to hypeR}

\item{top}{Limit number of genesets shown}
}
\value{
An igraph object
}
\description{
Do enrichment for all communities of the network
}
