% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{do.rwr}
\alias{do.rwr}
\title{Network propagation wrapper}
\usage{
do.rwr(ig, seeds, restart = 0.5)
}
\arguments{
\item{ig}{An igraph object}

\item{seeds}{One or more seed nodes}

\item{restart}{Restart probability between 0-1}
}
\value{
Propogation values
}
\description{
Network propagation wrapper
}
