% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{normalize.range}
\alias{normalize.range}
\title{Normalize values between a given range}
\usage{
normalize.range(x, a = 0, b = 1)
}
\arguments{
\item{x}{Values to normalize}

\item{a}{Min of range}

\item{b}{max of range}
}
\value{
Normalized values
}
\description{
Normalize values between a given range
}
